subscription_pack/requirements.txt
playwright==1.47.2
Pillow==10.3.0
a11ywatch-model==0.0.8
pyyaml==6.0.2
